distro/include: rework debug friendly optimization
The problem with DEBUG_BUILD is that we have a lot of debug information
in builds anyway and this is now unclear what it means. The
documentation.conf entry mentions packages which is now incorrect too.
The variable is a very old one and the interface is poorly designed
compared to other areas now.

This commit drop DEBUG_BUILD, use DEBUG_OPTIMIZE to instead, add
debug-optimize.bbclass and debug-optimize.inc. Use variable DEBUG_OPTIMIZE
to defer inherit debug-optimize.bbclass conditionally.

Make a config fragment core/yocto/debug-optimize to set DEBUG_OPTIMIZE ?= 1
to enable debug friendly optimize. For the recipe (such as qemu) which doesn't
work with debug optimization, set DEBUG_OPTIMIZE = "0" to disable it for a
given recipe even though config fragment core/yocto/debug-optimize is enabled

In bitbake.conf, use ??= to set *_OPTIMIZATION, in debug-optimize.inc
to use ?= to override *_OPTIMIZATION

In debug-optimize.bbclass, include_all debug-optimize.inc to allow other
layers to add their own debug optimization configurations.

User should use the following ways to enable debug optimize

  $ bitbake-config-build enable-fragment core/yocto/debug-optimize

Or

  $ echo 'OE_FRAGMENTS += "core/yocto/debug-optimize"' >> conf/local.conf

NOTE: we do not encourage user to enable debug friendly optimization by setting
DEBUG_OPTIMIZE = '1' directly, because the variable override is not certainty,
which is affected by the order of configuration parsing

Suggested-by: Peter Kjellerstedt <peter.kjellerstedt@axis.com>
Suggested-by: Richard Purdie <richard.purdie@linuxfoundation.org>
Signed-off-by: Hongxu Jia <hongxu.jia@windriver.com>
Signed-off-by: Mathieu Dubois-Briand <mathieu.dubois-briand@bootlin.com>